Make i18n tests work with a clean checkout
